Break the >900 lines __init__.py into modules
Tentative reorganization of the code in __init__.py into separate
modules.

Reason:
1) The code was getting really hard to navigate
2) We were exporting at the top level of the openchemistry package all
kinds on internal objects, such as 'private' functions, external imports
(ABC, jsonpath_rw, ...), making it hard to expose an curated public API
in jupyter lab's autocomplete.

Open to to discussion to see if we can come up with a better grouping.
