+++ openchemistry/recipes.py
@@ -1,0 +1,31 @@
+"""A collection of precanned workflows that can be achieved using the lower
+level openchemistry API
+"""
+
+class Reaction(object):
+    def __init__(self, equation):
+        self._equation = equation
+
+        (self._reactants, self._products) = equation.split('=>')
+        self._reactants = [x.strip() for x in self._reactants.split('+')]
+        self._products = [x.strip() for x in self._products.split('+')]
+
+    @property
+    def reactants(self):
+        return self._reactants
+
+    @property
+    def products(self):
+        return self._products
+
+    @property
+    def equation(self):
+        return '%s => %s' % (' + '.join(self.reactants), ' + '.join(self.products))
+
+def setup_reaction(equation):
+    return Reaction(equation)
+
+def compose_equation(equation, **vars):
+    equation = Environment(loader=BaseLoader()).from_string(equation)
+
+    return equation.render(**vars)
